Add support for <dependency name="" /> to link a library without copying files
git-svn-id: https://nekonme.googlecode.com/svn/trunk@2082 1509560c-5e2a-0410-865c-31c25e1cfdef

+++ tools/command-line/src/installers/IOSInstaller.hx
@@ -178,6 +178,16 @@ class IOSInstaller extends InstallerBase {
 			
 		}
 		
+		for (dependencyName in dependencyNames) {
+			
+			if (!StringTools.endsWith (dependencyName, ".framework")) {
+				
+				ndlls.push (new NDLL (dependencyName, null));
+				
+			}
+			
+		}
+		
 	}
 	
 	
@@ -226,20 +236,24 @@ class IOSInstaller extends InstallerBase {
 			
 			for (ndll in ndlls) {
 				
-				var releaseLib = ndll.getSourcePath ("iPhone", "lib" + ndll.name +  libExt);
-				var debugLib = ndll.getSourcePath ("iPhone", "lib" + ndll.name + "-debug" + libExt);
-				var releaseDest = projDestination + "lib/" + arch + "/lib" + ndll.name + ".a";
-				var debugDest = projDestination + "lib/" + arch + "-debug/lib" + ndll.name + ".a";
-				
-				FileHelper.copyIfNewer (releaseLib, releaseDest);
-				
-				if (FileSystem.exists (debugLib)) {
+				if (ndll.haxelib != null) {
 					
-					FileHelper.copyIfNewer (debugLib, debugDest);
+					var releaseLib = ndll.getSourcePath ("iPhone", "lib" + ndll.name +  libExt);
+					var debugLib = ndll.getSourcePath ("iPhone", "lib" + ndll.name + "-debug" + libExt);
+					var releaseDest = projDestination + "lib/" + arch + "/lib" + ndll.name + ".a";
+					var debugDest = projDestination + "lib/" + arch + "-debug/lib" + ndll.name + ".a";
 					
-				} else if (FileSystem.exists (debugDest)) {
+					FileHelper.copyIfNewer (releaseLib, releaseDest);
 					
-					FileSystem.deleteFile (debugDest);
+					if (FileSystem.exists (debugLib)) {
+						
+						FileHelper.copyIfNewer (debugLib, debugDest);
+						
+					} else if (FileSystem.exists (debugDest)) {
+						
+						FileSystem.deleteFile (debugDest);
+						
+					}
 					
 				}
